[ST20-bugfix/file-names] - rename category-services as category-service, rename ...next-auth as ...nextauth
