+++ stationery-next/app/api/auth/[...nextauth]/authOptions.ts
